pyproject.toml
[project]
name = "fastapi-diary"
version = "0.1.0"
description = "Diary fastapi 실습 project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["fastapi", "uvicorn", "python-dotenv", "tortoise-orm", "asyncpg", "aerich"]

[project.optional-dependencies]
dev = ["black", "ruff", "mypy", "pytest", "httpx"]
prod = ["gunicorn"]

[tool.aerich]
tortoise_orm = "app.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[dependency-groups]
dev = ["aerich>=0.9.1", "tomlkit>=0.13.3", "black", "ruff", "mypy", "pytest", "httpx", "isort"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["-v", "-s"]
asyncio_mode = "auto"
markers = [
    "anyio: marks tests as anyio tests"
]